refactor(guild): move guild runtime into its own package
