+++ src/guild/__init__.py
@@ -1,0 +1,1 @@
+"""Guild service package."""
